CSV

Documentation

configDir :: IO FilePath

readCSV :: FilePath -> IO [[Text]]

Reads CSV as 2-D T.Text list. If doesn't exist, creates new file with empty contents and returns []

readSingleColCSV :: FilePath -> IO [Text]

writeCSV :: FilePath -> [[Text]] -> IO ()

Write CSV from 2-D T.Text list

writeSingleColCSV :: FilePath -> [Text] -> IO ()

addToCSV :: FilePath -> [[Text]] -> IO ()

Appends the given tabular Text data to the CSV present at the given path. If no such CSV exists, a new one is created.

writeHashMapToCSV :: FilePath -> HashMap Text Text -> IO ()