Command.Parser

Description

Parser component for Commands. Its usage is mainly internal, to be used from Command.Command.

If you want to add your own parser argument type, you can either add it here (to be reused across multiple modules), or in your receive module (if it is only used there). Either works really, but sometimes you need the other to prevent looping imports.

Documentation

class ParsableArgument a where

A ParsableArgument is a dataclass that represents arguments that can be parsed from a message. Any datatype that is an instance of this dataclass can be used as function arguments for a command handler in command.

Methods

parserForArg :: Message -> Parser a

parserForArg msg returns a parser that contains the parsed element.

data RemainingText

Datatype wrapper for the remaining text in the input. Handy for capturing everything remaining. The accessor function getEm isn't really meant to be used since pattern matching can do everything. Open to renaming.

Example usage:

setStatus =
    command "status"
    $ \msg newStatus newType (Remaining newName) -> do
        ...

Constructors

Remaining

Fields

getEm :: Text

manyTill1 :: Stream s m t => ParsecT s u m a -> ParsecT s u m end -> ParsecT s u m [a]

manyTill1 p end is a parser that applies parser p one or more times until parser end succeeds. This is a variation on manyTill from Parsec.