Status

Documentation

editStatusFile :: UpdateStatusType -> ActivityType -> Text -> IO ()

editStatusFile puts the status values into "status.csv" by calling show on them and converting it to Text.

setStatusFromFile :: DiscordHandler ()

setStatusFromFile reads "status.csv" from the Config directory, and reads in the 3 columns as UpdateStatusType, ActivityType, and Text. The values are used to call updateStatus.

Incorrect formats (read parse errors) are ignored.

updateStatus :: UpdateStatusType -> ActivityType -> Text -> DiscordHandler ()

updateStatus updates the status through the Discord gateway. Therefore, it requires DiscordHandler and is not polymorphic.

Orphan instances

Read ActivityType

Instance details

Methods

readsPrec :: Int -> ReadS ActivityType

readList :: ReadS [ActivityType]

readPrec :: ReadPrec ActivityType

readListPrec :: ReadPrec [ActivityType]

Read UpdateStatusType
These datatypes in discord-haskell do not derive Read, but it's kinda necessary to do readMaybe, so here we go:

Instance details

Methods

readsPrec :: Int -> ReadS UpdateStatusType

readList :: ReadS [UpdateStatusType]

readPrec :: ReadPrec UpdateStatusType

readListPrec :: ReadPrec [UpdateStatusType]