Utils

Description

Documentation

emojiToUsableText :: Emoji -> Text

emojiToUsableText converts a given emoji to a text which can be used to display it in Discord.

sendMessageChan :: MonadDiscord m => ChannelId -> Text -> m ()

sendMessageChan attempts to send the given Text in the channel with the given channelID. Surpesses any error message(s), returning ().

sendReply :: MonadDiscord m => Message -> Bool -> Text -> m ()

sendReply attempts to send a reply to the given Message. Suppresses any error message(s), returning ().

sendMessageChanEmbed :: MonadDiscord m => ChannelId -> Text -> CreateEmbed -> m ()

sendMessageChanEmbed attempts to send the given embed with the given Text in the channel with the given channelID. Surpesses any error message(s), returning ().

sendMessageChanPingsDisabled :: MonadDiscord m => ChannelId -> Text -> m ()

sendMessageChanPingsDisabled acts in the same way as sendMessageChan, but disables all pings (everyone, user, @role) pings from the message.

sendMessageDM :: MonadDiscord m => UserId -> Text -> m ()

sendMessageDM attempts to send the given Text as a direct message to the user with the given UserId. Surpresses any error message(s), returning ().

sendFileChan :: (MonadDiscord m, MonadIO m) => ChannelId -> Text -> FilePath -> m ()

sendFileChan attempts to send the file at the provided FilePath in the channel with the provided ChannelId. The file attachment is annotated by the given Text. Surpresses any error message(s), returning ().

addReaction :: ChannelId -> MessageId -> Text -> DiscordHandler ()

addReaction attempts to add a reaction to the given message ID. Supresses any error message(s), returning ().

messageFromReaction :: MonadDiscord m => ReactionInfo -> m Message

messageFromReaction attempts to get the Message instance from a reaction.

pingUser :: User -> Text

pingUser constructs a minimal Text pinging the given user.

pingRole :: RoleId -> Text

pingRole constructs a minimal Text pinging the given role id.

pingAuthorOf :: Message -> Text

pingAuthorOf constructs a minimal Text pinging the author of a given message.

pingWithUsername :: Text -> GuildId -> DiscordHandler Text

pingWithUsername constructs a minimal Text pinging the the user with the given username from the given guild. On failure, returns an empty Text. On multiple such users, returns an empty Text.

stripAllPings :: Text -> Text

stripAllPings removes all pings from a given Text message.

newCommand

Arguments

:: Message
a message that needs to be handled

-> Text
the new command regex

-> ([Text] -> DiscordHandler ())
a function used to handle each message portion captured by the command regex

-> DiscordHandler ()
the over-all result of handling the message

newCommand should be used in the creation of a new Owen command. Given a Text command regex (lacking the : prefix and the trailing whitespace), along with a function that can handle the regex captures, the command can be used to create `Message -> DiscordHandler ()` message receivers.

newDevCommand :: Message -> Text -> ([Text] -> DiscordHandler ()) -> DiscordHandler ()

newDevCommand should be used in the creation of a new Owen dev command. Acts in the same way as newCommand, with the distinction that it constructs handlers that require the message author to be a developer. If they are not, the message author is messaged directly and reprimanded so harshly that they will never attempt to use a dev command ever again.

newModCommand :: Message -> Text -> ([Text] -> DiscordHandler ()) -> DiscordHandler ()

Similar to newDev command, however looks up the Moderator role name instead of using ID to determine the message authors role.

linkChannel :: ChannelId -> Text

linkChannel constructs a minimal Text linking the channel with the provided ID.

getMessageLink :: Message -> DiscordHandler Text

getMessageLink attempts to construct the Discord URL of the given message, as a Text.

hasRoleByName :: Message -> Text -> DiscordHandler Bool

hasRoleByName checks whether the provided message was sent by a user that has a role matching the provided Text exactly.

hasRoleByID :: Message -> RoleId -> DiscordHandler Bool

hasRoleByID checks whether the provided message was sent by a user that has a role matching the provided RoleId.

channelRequirement :: MonadDiscord m => String -> Message -> m (Maybe Text)

channelRequirement is a requirement for a Command to be in a certain channel.

isMod :: Message -> DiscordHandler Bool

isMod checks whether the provided message was sent by a user with the Moderator role.

devIDs :: FilePath

The FilePath to the configuration file listing OwenDev role IDs.

assetDir :: FilePath

The FilePath representing the location of the assets. TODO: Move to configDir

(=~=) :: Text -> Text -> Bool

The (=~=) function matches a given Text again a regex. Case-less in terms of owoifying.

getTimestampFromMessage :: Message -> Text

getTimestampFromMessages returns the given message's timestamp as Text, in the format `yyyy-mm-dd | hh:mm:ss`.

captureCommandOutput :: String -> IO Text

captureCommandOutput creates a new process from the desired command provided as a String. Then, it waits for the command to finish executing, returning its output as a Text.

update :: IO Bool

update calls a shell script that updates the bot's repo

snowflakeToInt :: Snowflake -> Integer

Converts Discord-Haskells Snowflake type to an integer

moveChannel :: GuildId -> ChannelId -> Int -> DiscordHandler ()

Moves channel position in guild

isEmojiValid :: Text -> GuildId -> DiscordHandler Bool

isEmojiValid determines whether an emoji (provided in Discord ::0-9 format) exists in the guild (or is a default emoji). Case insensitive.

isRoleInGuild :: MonadDiscord m => Text -> GuildId -> m (Maybe RoleId)

isRoleInGuild determines whether a role containing the given text exists in the guild (case insensitive). If it does, then it returns the role's ID. Otherwise, Nothing is returned.