Status

Documentation

updateStatus :: Text -> Text -> Text -> DiscordHandler ()

Convert intuitive strings into the respective DataTypes Passes values onto updateStatus' Although we revert to defaults if enums don't match, the caller of this function should always check first on their own and provide approriate error messages.

editStatusFile :: Text -> Text -> Text -> IO ()

setStatusFromFile :: DiscordHandler ()

Sets the status from file on bot launch Should be called only once.